requirements.txt
azure-iot-device==2.12.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
deprecation==2.1.0
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==3.4
janus==1.0.0
numpy==1.23.5
opencv-python==4.6.0.66
packaging==23.0
paho-mqtt==1.6.1
pandas==1.5.3
Pillow==9.4.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
requests-unixsocket==0.3.0
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.14
watchdog==2.2.1
imutils~=0.5.4